refactor(daemon): move send_file into FileService
send_file resolves the peer, adds the file to the blob store, and offers it over
FILES_ALPN, all FileService concerns. Move the logic onto FileService (using
registry peer resolution + transport blob store/endpoint); make guess_mime_type
and format_size pub(crate) so the service can reach them. MeshManager keeps a
thin delegate for the ray-mobile FFI + IPC dispatcher.

pair_with_device stays put for now: it auto-joins the primary's shared networks
via join_network, which is still on MeshManager. Build + clippy + 357 tests
green.
